+++ src/components/host-info/host-info.tsx
@@ -1,0 +1,38 @@
+interface HostInfoProps {
+  host: {
+    avatarUrl: string;
+    name: string;
+    isPro: boolean;
+  };
+  description: string;
+}
+
+function HostInfo({host, description}: HostInfoProps): JSX.Element {
+  return (
+    <div className="offer__host">
+      <h2 className="offer__host-title">Meet the host</h2>
+      <div className="offer__host-user user">
+        <div className="offer__avatar-wrapper offer__avatar-wrapper--pro user__avatar-wrapper">
+          <img
+            className="offer__avatar user__avatar"
+            src={host.avatarUrl}
+            width="74"
+            height="74"
+            alt="Host avatar"
+          />
+        </div>
+        <span className="offer__user-name">{host.name}</span>
+        {host.isPro ?
+          <span className="offer__user-status">Pro</span>
+          : null}
+      </div>
+      <div className="offer__description">
+        <p className="offer__text">
+          {description}
+        </p>
+      </div>
+    </div>
+  );
+}
+
+export default HostInfo;
