+++ src/pages/favorites-screen/favorites-screen.tsx
@@ -1,8 +1,7 @@
 import {Offer} from '../../types/offer';
-import {Link} from 'react-router-dom';
-import { AppRoute } from '../../const';
 import OffersList from '../../components/offers-list/offers-list';
-
+import Header from '../../components/header/header';
+import {useAppSelector} from '../../hooks/useAppSelector';
 type FavoritesScreenProps = {
   offers: Offer[];
 }
@@ -16,7 +15,10 @@ function FavoritesScreen({offers}: FavoritesScreenProps): JSX.Element {
     return result;
   }
 
+  const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
+  const user = useAppSelector((state) => state.user);
   const uniqueCities = getUniqueCities(offers);
+
   const classes = {
     article: 'favorites__card',
     img: 'favorites__image-wrapper',
@@ -28,35 +30,7 @@ function FavoritesScreen({offers}: FavoritesScreenProps): JSX.Element {
 
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to={AppRoute.Main}>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">3</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <Header authorizationStatus={authorizationStatus} user={user}/>
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
           <section className="favorites">
