+++ src/pages/offer-screen/offer-screen.tsx
@@ -6,9 +6,10 @@ import { fetchRoomAction, fetchNearByHotelsAction, fetchCommentsAction } from '.
 import capitalizeFirstLetter from '../../helper-functions';
 import Feedbacks from '../../components/feedbacks/feedbacks';
 import Header from '../../components/header/header';
-import OffersList from '../../components/offers-list/offers-list';
 import Map from '../../components/map/map';
 import Spinner from '../../components/spinner/spinner';
+import NearPlaces from '../../components/near-places/near-places';
+import HostInfo from '../../components/host-info/host-info';
 
 function OfferScreen(): JSX.Element {
   const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
@@ -40,13 +41,6 @@ function OfferScreen(): JSX.Element {
     );
   }
 
-  const classes = {
-    article: 'near-places__card',
-    img: 'near-places__image-wrapper',
-    info: '',
-    isPremiumBlockShow: false
-  };
-
   if (!id || !offer) {
     return <Navigate to="/404" replace />;
   }
@@ -57,7 +51,6 @@ function OfferScreen(): JSX.Element {
   return (
     <div className="page">
       <Header authorizationStatus={authorizationStatus} user={user}/>
-
       <main className="page__main page__main--offer">
         <section className="offer">
           <div className="offer__gallery-container container">
@@ -127,29 +120,7 @@ function OfferScreen(): JSX.Element {
                   })}
                 </ul>
               </div>
-              <div className="offer__host">
-                <h2 className="offer__host-title">Meet the host</h2>
-                <div className="offer__host-user user">
-                  <div className="offer__avatar-wrapper offer__avatar-wrapper--pro user__avatar-wrapper">
-                    <img
-                      className="offer__avatar user__avatar"
-                      src={host.avatarUrl}
-                      width="74"
-                      height="74"
-                      alt="Host avatar"
-                    />
-                  </div>
-                  <span className="offer__user-name">{host.name}</span>
-                  {host.isPro ?
-                    <span className="offer__user-status">Pro</span>
-                    : null}
-                </div>
-                <div className="offer__description">
-                  <p className="offer__text">
-                    {description}
-                  </p>
-                </div>
-              </div>
+              <HostInfo host={host} description={description}/>
               <Feedbacks feedbacks={feedbacks} hotelId={id} authorizationStatus={authorizationStatus}/>
             </div>
           </div>
@@ -161,14 +132,7 @@ function OfferScreen(): JSX.Element {
           />
         </section>
         <div className="container">
-          <section className="near-places places">
-            <h2 className="near-places__title">
-                  Other places in the neighbourhood
-            </h2>
-            <div className="near-places__list places__list">
-              <OffersList offers={ nearestRooms } className={classes}/>
-            </div>
-          </section>
+          <NearPlaces nearestRooms={nearestRooms}/>
         </div>
       </main>
     </div>
