+++ src/store/action.ts
@@ -11,10 +11,6 @@ export const setOffersList = createAction('offers/setOffersList', (value: City) 
   payload: value,
 }));
 
-export const setSortedOffersList = createAction('offers/setSortedOffersList', (value: Offer[]) => ({
-  payload: value,
-}));
-
 export const loadOffers = createAction<Offer[]>('data/loadOffers');
 
 export const loadRoom = createAction<Offer>('data/loadRoom');
